% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{greenwood}
\alias{greenwood}
\title{The text of Thomas Hardy's novel "Under the Greenwood Tree"}
\format{A character vector with 7003 elements}
\source{
\url{http://www.gutenberg.org/ebooks/2662}
}
\usage{
greenwood
}
\description{
A dataset containing the text of Thomas Hardy's novel "Under the Greenwood Tree",
published in 1872.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
