% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{jude}
\alias{jude}
\title{The text of Thomas Hardy's novel "Jude the Obscure"}
\format{
A character vector with 15848 elements
}
\source{
\url{http://www.gutenberg.org/ebooks/153}
}
\usage{
jude
}
\description{
A dataset containing the text of Thomas Hardy's novel "Jude the Obscure",
published in 1895.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
