% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{mayor}
\alias{mayor}
\title{The text of Thomas Hardy's novel "The Mayor of Casterbridge"}
\format{A character vector with 12464 elements}
\source{
\url{http://www.gutenberg.org/ebooks/143}
}
\usage{
mayor
}
\description{
A dataset containing the text of Thomas Hardy's 1886 novel "The Mayor of
Casterbridge". The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}

