% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{native}
\alias{native}
\title{The text of Thomas Hardy's novel "The Return of the Native"}
\format{A character vector with 15021 elements}
\source{
\url{http://www.gutenberg.org/ebooks/122}
}
\usage{
native
}
\description{
A dataset containing the text of Thomas Hardy's 1878 novel "The Return of 
the Native". The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}

