% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{tess}
\alias{tess}
\title{The text of Thomas Hardy's novel "Tess of the D'Urbervilles"}
\format{
A character vector with 16091 elements
}
\source{
\url{http://www.gutenberg.org/ebooks/110}
}
\usage{
tess
}
\description{
A dataset containing the text of Thomas Hardy's novel "Tess of the
D'Urbervilles" published in 1891. The UTF-8 plain text was sourced from
Project Gutenberg and is divided into elements of up to about 70 characters
each. (Some elements are blank.)
}
\keyword{datasets}
