% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{woodlanders}
\alias{woodlanders}
\title{The text of Thomas Hardy's novel "The Woodlanders"}
\format{A character vector with 14095 elements}
\source{
\url{http://www.gutenberg.org/ebooks/482}
}
\usage{
woodlanders
}
\description{
A dataset containing the text of Thomas Hardy's 1887 novel "The Woodlanders".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each. 
(Some elements are blank.)
}
\keyword{datasets}

